处理极光推送的标签设置，打开的时候向后台请求标签，登录的时候向后台请求标签，推出登录的时候向后台请求标签

+++ app/src/main/java/com/yslc/util/FileUtil.java
@@ -92,6 +92,9 @@ public class FileUtil {
     public static void deleteAllFile(String filePath) {
         File file = new File(getSdCardPath() + File.separator + filePath);
         File[] fileArr = file.listFiles();
+        if(fileArr == null){
+            return;
+        }
         for (File files : fileArr) {
             files.delete();
         }
